requirements.txt
streamlit
PyPDF2
python-docx
mistralai
langchain
python-dotenv
openai
python-dotenv
django
gunicorn
pillow
gunicorn==21.2.0
psycopg2-binary
django-cors-headers==4.6.0
qrcode==7.4.2
sqlparse==0.5.1
virtualenv==20.26.3
requests